requirements.txt
absl-py==0.10.0
appnope==0.1.0
argon2-cffi==20.1.0
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
future==0.18.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
html2text==2020.1.16
html5lib==1.1
idna==2.9
importlib-metadata==1.7.0
investpy==1.0
ipdb==0.13.4
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==5.5.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.6.2
Markdown==3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mock==4.0.3
mpmath==1.1.0
multitasking==0.0.9
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
nose==1.3.7
notebook==6.1.4
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prettytable==1.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
pyecharts==1.7.1
Pygments==2.7.1
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
ruamel-yaml==0.15.87
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.22
sympy==1.6.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-hub==0.9.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typing-extensions==3.7.4.3
Unidecode==1.1.2
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.3.0
